requirement.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
future==0.16.0
get==1.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
post==1.0.1
pprint==0.1
public==1.0.1
pyTelegramBotAPI==3.6.2
python-telegram-bot==10.0.2
query-string==1.0.1
request==1.0.1
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
